Component Structure Change Homepage
