+++ client/src/Styles/HomePage.css
@@ -8,7 +8,7 @@
 /*-------------------------------- .homepage-section-1-------------------------------- */
 .homepage-section-1 {
   width: 100%;
-  height: 85vh;
+  height: 80vmin;
   display: flex;
   justify-content: center;
   align-content: center;
