+++ client/src/components/HomePage_Sections/HP_Section3.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+const Section3 = () => {
+  return (
+    <section className='homepage-section-3'>
+      <div className='hp-section3-griditem1'>
+        <h2> WHY WAIT....? GRAB YOUR SNEAKERS NOW. </h2>
+      </div>
+      <div className='hp-section3-griditem2'></div>
+    </section>
+  );
+};
+
+export default Section3;
